backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0
requests==2.32.3
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
# Alternative OCR library if tesseract fails
easyocr==1.7.0
